requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.6
django-admin-interface==0.12.2
django-colorfield==0.3.1
django-crispy-forms==1.9.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.1.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.2
psycopg2-binary==2.8.5
pytz==2020.1
PyYAML==5.3.1
sklearn==0.0
sqlparse==0.3.1
tablib==1.1.0
tensorboard==2.1.1
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
